// src/interfaces/news.ts
export interface Article {
  id: number;
  title: string;
  url: string;
  image_url: string | null;
  news_site: string;
  summary: string;
  published_at: string;
  updated_at: string;
  featured: boolean;
}

export interface NewsFormData {
  title: string;
  summary: string;
  url: string;
  image_url: string;
  news_site: string;
}

export interface NewsState {
  articles: Article[];
  loading: boolean;
  error: string | null;
  totalCount: number;
}

export interface ApiResponse {
  count: number;
  next: string | null;
  previous: string | null;
  results: Article[];
}